+++ nmdc_runtime/site/drsobjects/registration.py
@@ -2,7 +2,6 @@ import json
 import os
 import re
 from datetime import datetime, timezone, timedelta
-from functools import lru_cache
 from pathlib import Path
 from tempfile import TemporaryDirectory
 
@@ -11,9 +10,9 @@ from bs4 import BeautifulSoup
 
 from nmdc_runtime.api.models.object import DrsObjectIn
 from nmdc_runtime.util import (
-    nmdc_jsonschema,
     drs_metadata_for,
     nmdc_jsonschema_validator,
+    specialize_activity_set_docs,
 )
 
 pattern = re.compile(r"https?://(?P<domain>[^/]+)/(?P<path>.+)")
@@ -87,43 +86,6 @@ def drs_object_in_for(url):
 def create_drs_object_for(url, drs_object_in, client):
     rv = client.create_object(json.loads(drs_object_in.json(exclude_unset=True)))
     return {"url": url, "response": rv}
-
-
-@lru_cache
-def get_type_collections():
-    return {
-        f'nmdc:{spec["items"]["$ref"].split("/")[-1]}': collection_name
-        for collection_name, spec in nmdc_jsonschema["properties"].items()
-        if collection_name.endswith("_set")
-    }
-
-
-def specialize_activity_set_docs(docs):
-    validation_errors = {}
-    type_collections = get_type_collections()
-    if "activity_set" in docs:
-        for doc in docs["activity_set"]:
-            doc_type = doc["type"]
-            try:
-                collection_name = type_collections[doc_type]
-            except KeyError:
-                msg = (
-                    f"activity_set doc {doc.get('id', '<id missing>')} "
-                    f"has type {doc_type}, which is not in NMDC Schema. "
-                    "Note: Case is sensitive."
-                )
-                if "activity_set" in validation_errors:
-                    validation_errors["activity_set"].append(msg)
-                else:
-                    validation_errors["activity_set"] = [msg]
-                continue
-
-            if collection_name in docs:
-                docs[collection_name].append(doc)
-            else:
-                docs[collection_name] = [doc]
-        del docs["activity_set"]
-    return docs, validation_errors
 
 
 def validate_as_metadata_and_ensure_tags_for(
