+++ nmdc_runtime/util.py
@@ -3,20 +3,35 @@ import mimetypes
 import os
 import pkgutil
 from collections.abc import Iterable
+from contextlib import AbstractContextManager
 from copy import deepcopy
 from datetime import datetime, timezone
 from functools import lru_cache
 from io import BytesIO
 from pathlib import Path
+from uuid import uuid4
 
 import fastjsonschema
 import requests
 from frozendict import frozendict
+from jsonschema.validators import Draft7Validator
 from nmdc_schema.get_nmdc_view import ViewGetter
-from toolz import merge, pluck
+from pydantic import conint, BaseModel
+from pymongo.database import Database as MongoDatabase
+from pymongo.errors import OperationFailure
+from toolz import merge, unique
 
 from nmdc_runtime.api.core.util import sha256hash_from_file
 from nmdc_runtime.api.models.object import DrsObjectIn
+
+
+@lru_cache
+def get_type_collections():
+    return {
+        f'nmdc:{spec["items"]["$ref"].split("/")[-1]}': collection_name
+        for collection_name, spec in nmdc_jsonschema["properties"].items()
+        if collection_name.endswith("_set")
+    }
 
 
 def without_id_patterns(nmdc_jsonschema):
@@ -215,3 +230,202 @@ def nmdc_activity_collection_names():
         if rng in acts:
             slots.append(slot)
     return slots
+
+
+def all_docs_have_unique_id(coll) -> bool:
+    first_doc = coll.find_one({}, ["id"])
+    if first_doc is None or "id" not in first_doc:
+        # short-circuit exit for empty collection or large collection via first-doc peek.
+        return False
+
+    total_count = coll.count_documents({})
+    return len(coll.distinct("id")) == total_count
+
+
+def specialize_activity_set_docs(docs):
+    validation_errors = {}
+    type_collections = get_type_collections()
+    if "activity_set" in docs:
+        for doc in docs["activity_set"]:
+            doc_type = doc["type"]
+            try:
+                collection_name = type_collections[doc_type]
+            except KeyError:
+                msg = (
+                    f"activity_set doc {doc.get('id', '<id missing>')} "
+                    f"has type {doc_type}, which is not in NMDC Schema. "
+                    "Note: Case is sensitive."
+                )
+                if "activity_set" in validation_errors:
+                    validation_errors["activity_set"].append(msg)
+                else:
+                    validation_errors["activity_set"] = [msg]
+                continue
+
+            if collection_name in docs:
+                docs[collection_name].append(doc)
+            else:
+                docs[collection_name] = [doc]
+        del docs["activity_set"]
+    return docs, validation_errors
+
+
+collection_name_to_class_name = {
+    db_prop: db_prop_spec["items"]["$ref"].split("/")[-1]
+    for db_prop, db_prop_spec in get_nmdc_jsonschema_dict()["$defs"]["Database"][
+        "properties"
+    ].items()
+    if "items" in db_prop_spec and "$ref" in db_prop_spec["items"]
+}
+
+
+@lru_cache
+def schema_collection_names_with_id_field():
+    return {
+        coll_name
+        for coll_name, class_name in collection_name_to_class_name.items()
+        if "id" in get_nmdc_jsonschema_dict()["$defs"][class_name].get("properties", {})
+    }
+
+
+def ensure_unique_id_indexes(mdb: MongoDatabase):
+    """Ensure that any collections with an "id" field have an index on "id"."""
+    candidate_names = (
+        set(mdb.list_collection_names()) | schema_collection_names_with_id_field()
+    )
+    for collection_name in candidate_names:
+        if collection_name.startswith("system."):  # reserved by mongodb
+            continue
+
+        if (
+            collection_name in schema_collection_names_with_id_field()
+            or all_docs_have_unique_id(mdb[collection_name])
+        ):
+            mdb[collection_name].create_index("id", unique=True)
+
+
+class UpdateStatement(BaseModel):
+    q: dict
+    u: dict
+    upsert: bool = False
+    multi: bool = False
+
+
+class DeleteStatement(BaseModel):
+    q: dict
+    limit: conint(ge=0, le=1) = 1
+
+
+class OverlayDBError(Exception):
+    pass
+
+
+class OverlayDB(AbstractContextManager):
+    """Provides a context whereby a base Database is overlaid with a temporary one.
+
+    If you need to run basic simulations of updates to a base database,
+    you don't want to actually commit transactions to the base database.
+
+    For example, to insert or replace (matching on "id") many documents into a collection in order
+    to then validate the resulting total set of collection documents, an OverlayDB writes to
+    an overlay collection that "shadows" the base collection during a "find" query
+    (the "merge_find" method of an OverlayDB object): if a document with `id0` is found in the
+    overlay collection, that id is marked as "seen" and will not also be returned when
+    subsequently scanning the (unmodified) base-database collection.
+
+    Mongo "update" commands (as the "apply_updates" method) are simulated by first copying affected
+    documents from a base collection to the overlay, and then applying the updates to the overlay,
+    so that again, base collections are unmodified, and a "merge_find" call will produce a result
+    *as if* the base collection(s) were modified.
+
+    Mongo deletions (as the "delete" method) also copy affected documents from the base collection
+    to the overlay collection, and flag them using the "_deleted" field. In this way, a `merge_find`
+    call will match a relevant document given a suitable filter, and will mark the document's id
+    as "seen" *without* returning the document. Thus, the result is as if the document were deleted.
+
+    Usage:
+    ````
+    with OverlayDB(mdb) as odb:
+        # do stuff, e.g. `odb.replace_or_insert_many(...)`
+    ```
+    """
+
+    def __init__(self, mdb: MongoDatabase):
+        self._bottom_db = mdb
+        self._top_db = self._bottom_db.client.get_database(f"overlay-{uuid4()}")
+        ensure_unique_id_indexes(self._top_db)
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self, exc_type, exc_value, traceback):
+        self._bottom_db.client.drop_database(self._top_db.name)
+
+    def replace_or_insert_many(self, coll_name, documents: list):
+        try:
+            self._top_db[coll_name].insert_many(documents)
+        except OperationFailure as e:
+            raise OverlayDBError(str(e.details))
+
+    def apply_updates(self, coll_name, updates: list):
+        """prepare overlay db and apply updates to it."""
+        assert all(UpdateStatement(**us) for us in updates)
+        for update_spec in updates:
+            for bottom_doc in self._bottom_db[coll_name].find(update_spec["q"]):
+                self._top_db[coll_name].insert_one(bottom_doc)
+        try:
+            self._top_db.command({"update": coll_name, "updates": updates})
+        except OperationFailure as e:
+            raise OverlayDBError(str(e.details))
+
+    def delete(self, coll_name, deletes: list):
+        """ "apply" delete command by flagging docs in overlay database"""
+        assert all(DeleteStatement(**us) for us in deletes)
+        for delete_spec in deletes:
+            for bottom_doc in self._bottom_db[coll_name].find(
+                delete_spec["q"], limit=delete_spec.get("limit", 1)
+            ):
+                bottom_doc["_deleted"] = True
+                self._top_db[coll_name].insert_one(bottom_doc)
+
+    def merge_find(self, coll_name, find_spec: dict):
+        """Yield docs first from overlay and then from base db, minding deletion flags."""
+        # ensure projection of "id" and "_deleted"
+        if "projection" in find_spec:
+            proj = find_spec["projection"]
+            if isinstance(proj, dict):
+                proj = merge(proj, {"id": 1, "_deleted": 1})
+            elif isinstance(proj, list):
+                proj = list(unique(proj + ["id", "_deleted"]))
+
+        top_docs = self._top_db[coll_name].find(**find_spec)
+        bottom_docs = self._bottom_db[coll_name].find(**find_spec)
+        top_seen_ids = set()
+        for doc in top_docs:
+            if not doc.get("_deleted"):
+                yield doc
+            top_seen_ids.add(doc["id"])
+
+        for doc in bottom_docs:
+            if doc["id"] not in top_seen_ids:
+                yield doc
+
+
+def validate_json(docs: dict, mdb: MongoDatabase):
+    validator = Draft7Validator(get_nmdc_jsonschema_dict())
+    docs = deepcopy(docs)
+    docs, validation_errors = specialize_activity_set_docs(docs)
+
+    for coll_name, coll_docs in docs.items():
+        errors = list(validator.iter_errors({coll_name: coll_docs}))
+        validation_errors[coll_name] = [e.message for e in errors]
+        try:
+            with OverlayDB(mdb) as odb:
+                odb.replace_or_insert_many(coll_name, coll_docs)
+        except OverlayDBError as e:
+            validation_errors[coll_name].append(str(e))
+
+    if all(len(v) == 0 for v in validation_errors.values()):
+        return {"result": "All Okay!"}
+    else:
+        return {"result": "errors", "detail": validation_errors}
